requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiofiles==23.2.1
    # via quart
black==23.3.0
    # via -r requirements.in
blinker==1.5
    # via quart
classify-imports==4.2.0
    # via reorder-python-imports
click==8.1.3
    # via
    #   black
    #   quart
evdev==1.6.1
    # via -r requirements.in
exceptiongroup==1.1.1
    # via pytest
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.14.4
    # via quart
hyperframe==6.0.1
    # via h2
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via quart
jinja2==3.1.2
    # via quart
markupsafe==2.1.3
    # via
    #   jinja2
    #   quart
    #   werkzeug
mouseinfo==0.1.3
    # via pyautogui
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via
    #   black
    #   pytest
pathspec==0.11.1
    # via black
platformdirs==3.2.0
    # via black
pluggy==1.0.0
    # via pytest
priority==2.0.0
    # via hypercorn
pyautogui==0.9.53
    # via -r requirements.in
pygetwindow==0.0.9
    # via pyautogui
pymsgbox==1.0.9
    # via pyautogui
pyperclip==1.8.2
    # via mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==0.1.28
    # via pyautogui
pytest==7.3.0
    # via -r requirements.in
python3-xlib==0.15
    # via
    #   mouseinfo
    #   pyautogui
pytweening==1.0.4
    # via pyautogui
pyyaml==6.0
    # via -r requirements.in
quart==0.18.4
    # via -r requirements.in
reorder-python-imports==3.9.0
    # via -r requirements.in
tomli==2.0.1
    # via
    #   black
    #   hypercorn
    #   pytest
werkzeug==2.3.7
    # via quart
wsproto==1.2.0
    # via hypercorn
